core
updated

+++ ServerLauncher/Form1.cs
@@ -1,22 +1,4 @@
-﻿/*
- * Copyright (C) 2013 APS
- *	http://AllPrivateServer.com
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- */
- 
+﻿
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -44,67 +26,74 @@ namespace ServerLauncher
 
         private void B_start_Click(object sender, EventArgs e)
         {
-            AccountCacher = new Process();
-            AccountCacher.StartInfo.FileName = "AccountCacher.exe";
-            AccountCacher.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            AccountCacher.Start();
+            B_start.Enabled = false;
 
-            Thread.Sleep(500);
+            if (StartAccountCheckBox.Checked)
+            {
+                AccountCacher = new Process();
+                AccountCacher.StartInfo.FileName = "AccountCacher.exe";
+                AccountCacher.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                AccountCacher.Start();
 
-            LauncherServer = new Process();
-            LauncherServer.StartInfo.FileName = "LauncherServer.exe";
-            LauncherServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            LauncherServer.Start();
+                Thread.Sleep(500);
+            }
 
-            Thread.Sleep(500);
+            if (StartLauncherCheckBox.Checked)
+            {
+                LauncherServer = new Process();
+                LauncherServer.StartInfo.FileName = "LauncherServer.exe";
+                LauncherServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                LauncherServer.Start();
 
-            LobbyServer = new Process();
-            LobbyServer.StartInfo.FileName = "LobbyServer.exe";
-            LobbyServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            LobbyServer.Start();
+                Thread.Sleep(500);
+            }
 
-            Thread.Sleep(500);
+            if (StartLobbyCheckBox.Checked)
+            {
+                LobbyServer = new Process();
+                LobbyServer.StartInfo.FileName = "LobbyServer.exe";
+                LobbyServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                LobbyServer.Start();
 
-            WorldServer = new Process();
-            WorldServer.StartInfo.FileName = "WorldServer.exe";
-            WorldServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            WorldServer.Start();
+                Thread.Sleep(500);
+            }
+
+            if (StartWorldCheckBox.Checked)
+            {
+                WorldServer = new Process();
+                WorldServer.StartInfo.FileName = "WorldServer.exe";
+                WorldServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                WorldServer.Start();
+            }
+
+            B_start.Enabled = true;
         }
 
         private void B_stop_Click(object sender, EventArgs e)
         {
             try
             {
-                WorldServer.Kill();
+                if( WorldServer != null ) WorldServer.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
+            
             try
             {
-                LobbyServer.Kill();
+                if (LobbyServer != null) LobbyServer.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
+            
             try
             {
-                LauncherServer.Kill();
+                if (LauncherServer != null) LauncherServer.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
+            
             try
             {
-                AccountCacher.Kill();
+                if (AccountCacher != null) AccountCacher.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
         }
     }
 }
